refactor(shared): isolate hub resource types and tidy generate core imports and specs
Move ResourceType, RESOURCE_TYPES and isValidResourceType out of
calm-hub-client.ts (which imports axios) into a new resource-types.ts, and
have document-id-utils.ts import from there instead. calm-hub-client.ts
re-exports the moved names so its public surface and the root barrel are
unchanged. This keeps the axios-based hub client out of the browser
bundle's graph, since document-id-utils.ts is imported by browser.ts.

Add the missing .js suffixes to generate-core.ts's local imports for
consistency with its other imports.

Move the unmocked 'generate core' describe block out of generate.spec.ts
into a new generate-core.spec.ts that imports generate-core directly and
never mocks anything, so it needs no vi.doUnmock/resetModules dance that
only worked because it ran last in the file.
